Relocate SearchViewVisibleSuggestionTelemetryInfo
